renames and rect corners round fix
Scissor cut the outline, so the rounded corners of the input text were not visible.
SOLVED: the size of the inner area was reduced by the thickness of the outline.
Added output text for outputting messages.
Added a helper class based on a ring buffer for an infinite stream of text messages in output text.
Renamed old macros, functions and constants